moved user stuff to own file
