--- backend/api_structs.go
@@ -1,6 +1,0 @@
-package main
-
-type AddUser struct {
-	Name string `json:"name,omitempty"`
-	Pswd string `json:"pswd,omitempty"`
-}
