--- backend/db_structs.go
@@ -1,7 +1,0 @@
-package main
-
-// A user struct containing the public info of a user
-type PubUser struct {
-	Id   int
-	Name string
-}
